refactor(main): make regular venv persistent
Changes the behavior of the previously fixed regular venv. Now it is
more persistent, meaning that it has higher priority than pytomata's
pyenv venv. If user has regular venv on and goes to the directory which
contains information about the pytomata's pyenv venv, pytomata does not
take precedence. It effectively means that if the user uses a regular
venv, they will have to turn it off manually by the `deactivate` command.